+++ frontend/src/router/index.tsx
@@ -9,9 +9,9 @@ import CategoryPage from "../components/Categories/Categories";
 import SubcategoryItems from "../components/SubCategories/Subcategories";
 import GetAllReviews from "../components/Reviews/getAllReviews";
 import Cart from "../components/Cart/Cart";
-import Checkout from "../components/CheckoutSuccess.css/Checkout";
-import CheckoutForm from "../components/CheckoutSuccess.css/CheckoutForm";
-import CheckoutSuccess from "../components/CheckoutSuccess.css/CheckoutSuccess";
+import Checkout from "../components/Checkout/Checkout";
+import CheckoutForm from "../components/Checkout/CheckoutForm";
+import CheckoutSuccess from "../components/Checkout/CheckoutSuccess";
 
 export const router = createBrowserRouter([
   {
